add subscription logic and small changes in thymeleaf templates

+++ src/main/java/regexit/feeder/repos/MessageRepo.java
@@ -5,6 +5,6 @@ import regexit.feeder.domain.Message;
 
 import java.util.List;
 
-public interface MessageRepo extends JpaRepository<Message, Integer> {
+public interface MessageRepo extends JpaRepository<Message, Long> {
     List<Message> findByTag(String tag);
 }
